+++ src/onegov.swissvotes/onegov/swissvotes/tests/test_views.py
@@ -1,0 +1,78 @@
+from onegov.core.utils import module_path
+from pytest import mark
+from webtest import TestApp as Client
+from webtest.forms import Upload
+from onegov.swissvotes.models import SwissVote
+
+
+def login(client):
+    login = client.get('/auth/login')
+    login.form['username'] = 'admin@example.org'
+    login.form['password'] = 'hunter2'
+    login.form.submit()
+
+
+@mark.parametrize("file", [
+    module_path('onegov.swissvotes', 'tests/fixtures/votes.xlsx'),
+])
+def test_update_votes(swissvotes_app, file):
+    client = Client(swissvotes_app)
+    client.get('/locale/de_CH').follow()
+
+    login(client)
+
+    with open(file, 'rb') as f:
+        content = f.read()
+
+    # todo: fixture kleiner!
+
+    # Upload
+    manage = client.get('/votes/update')
+    manage.form['dataset'] = Upload(
+        'votes.xlsx',
+        content,
+        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+    )
+    manage = manage.form.submit().follow()
+    assert "Datensatz aktualisiert (634 hinzugefügt, 0 geändert)" in manage
+
+    session = swissvotes_app.session()
+    vote = session.query(SwissVote).filter_by(bfs_number=82.2).one()
+    assert str(vote.bfs_number) == '82.20'
+    assert vote.date.isoformat() == '1920-03-21'
+    assert vote.legislation_number == 25
+    assert vote.decade.lower == 1920
+    assert vote.decade.upper == 1929
+    assert vote.title == (
+        "Gegenentwurf zur Volksinitiative "
+        "«für ein Verbot der Errichtung von Spielbanken»"
+    )
+    assert [str(pa) for pa in vote.policy_areas] == ['4.41.413', '4.44.443']
+    assert str(vote.result_turnout) == '60.2323410000'
+    assert vote.recommendations_parties['Nay'][0].name == 'sps'
+
+    # Upload (unchanged)
+    manage = client.get('/votes/update')
+    manage.form['dataset'] = Upload(
+        'votes.xlsx',
+        content,
+        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+    )
+    manage = manage.form.submit().follow()
+    assert "Datensatz aktualisiert (0 hinzugefügt, 0 geändert)" in manage
+
+    # Download
+    csv = client.get('/votes/csv').body
+    xlsx = client.get('/votes/xlsx').body
+
+    # Upload (roundtrip)
+    manage = client.get('/votes/update')
+    manage.form['dataset'] = Upload(
+        'votes.xlsx',
+        xlsx,
+        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+    )
+    manage = manage.form.submit().follow()
+    assert "Datensatz aktualisiert (0 hinzugefügt, 0 geändert)" in manage
+
+    assert csv == client.get('/votes/csv').body